azure-ai/scripts/requirements.txt
azure-search-documents==11.6.0
